+++ Runtime/CustomAttributes/HideInArtificeAttribute.cs
@@ -1,0 +1,6 @@
+namespace ArtificeToolkit.Attributes
+{
+    public class HideInArtificeAttribute : CustomAttribute
+    {
+    }
+}